inserción datos de prueba para la escritura del archivo de transacciones

+++ ProgramaAuxiliar/src/modelos/Transaccion.java
@@ -4,11 +4,11 @@ package modelos;
 public class Transaccion {
     
     private int numTelefoNroCuenta;
-    private int monto;
+    private double monto;
     private String Descripcion;
     private String tipoTransaccion;
 
-    public Transaccion(int numTelefoNroCuenta, int monto, String Descripcion, String tipoTransaccion) {
+    public Transaccion(int numTelefoNroCuenta, double monto, String Descripcion, String tipoTransaccion) {
         this.numTelefoNroCuenta = numTelefoNroCuenta;
         this.monto = -monto;
         this.Descripcion = Descripcion;
